+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.20
+* Fix `getBluetoothAvailabilityState` reporting wrong Bluetooth status on browsers where Web Bluetooth can be globally disabled
+* `startScan` API on Web platforms now throws a `WebBluetoothGloballyDisabled` exception if scanning cannot proceed due to Web Bluetooth being globally disabled
+
 ## 0.19.0
 * Add new high level API
 * Deprecated: use `read` instead of `readValue`